fix(productivity-suite): ensure that fragment q:base paths are absolute
If you do not do this, then the "fragment base path" is appended to
the current URL.
This can result in Qwik downloading two copies of the same ESM file
and so there are two copies of the module level global variables.
In particular the _context variable which is used by Qwik to determine
the current context in which to run hooks.
If that happens then you might see "code-14" errors where Qwik thinks
that the context has not been set before triggering code that needs a
lexical scope to be set.