fix(productivity-suite): ensure that we only run side-effects once per fragment (#86)
Previously, if an outlet was added and then removed quickly, before the
fragment had completed its execution, we still added it to the set of
previously unmounted fragments. This meant that when the outlet got added
a second time, we would manually run the side-effects, even though they
would still be automatically run as part of the first-time load.

Now we only add fragments to the previously unmounted set if it completed
its execution.